Move components that render pages to "views" -folder
